chore: rework folder structure
