--- core/threshold-networking/src/grpc.rs
@@ -1,1501 +1,0 @@
-//! gRPC-based networking.
-
-use super::ggen::gnetworking_server::{Gnetworking, GnetworkingServer};
-use super::ggen::{HealthCheckRequest, HealthCheckResponse, SendValueRequest, SendValueResponse};
-use super::sending_service::{
-    GrpcSendingService, NetworkSession, SendingService, now_activity_millis,
-};
-use super::tls::extract_subject_from_cert;
-use crate::constants::{
-    DISCARD_INACTIVE_SESSION_INTERVAL_SECS, INITIAL_INTERVAL_MS, MAX_BUFFERED_FUTURE_MSGS,
-    MAX_ELAPSED_TIME, MAX_EN_DECODE_MESSAGE_SIZE, MAX_INTERVAL,
-    MAX_OPENED_INACTIVE_SESSIONS_PER_PARTY, MAX_WAITING_TIME_MESSAGE_QUEUE, MESSAGE_LIMIT,
-    MULTIPLIER, NETWORK_TIMEOUT_BK, NETWORK_TIMEOUT_BK_SNS, NETWORK_TIMEOUT_LONG,
-    SESSION_CLEANUP_INTERVAL_SECS, SESSION_STATUS_UPDATE_INTERVAL_SECS,
-};
-use crate::ggen::Status;
-use crate::health_check::HealthCheckSession;
-use async_trait::async_trait;
-use dashmap::DashMap;
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-use std::sync::atomic::{AtomicU64, Ordering};
-use std::sync::{Arc, LazyLock, Weak};
-use threshold_types::role::{RoleKind, RoleTrait};
-use threshold_types::session_id::SessionId;
-use threshold_types::{
-    network::{NetworkMode, Networking},
-    party::{MpcIdentity, RoleAssignment},
-};
-use tokio::sync::{
-    Mutex,
-    mpsc::{Receiver, Sender, channel},
-};
-use tokio::time::{Duration, Instant};
-
-use tonic::transport::CertificateDer;
-use tonic::transport::server::TcpConnectInfo;
-use x509_parser::parse_x509_certificate;
-
-/// Network configuration for core-to-core communication.
-///
-/// Every field is optional: an absent field falls back to the corresponding
-/// default in [`crate::constants`] via the `get_*` accessors below. This means
-/// a fully-absent config (`CoreToCoreNetworkConfig::default()`, all `None`) is
-/// equivalent to "use all defaults", and a partial config overrides only the
-/// fields it sets. Callers should always read through the `get_*` accessors
-/// rather than the raw fields so the defaulting stays in one place.
-///
-/// WARNING: this may be printed for debugging and hence should NOT contain any secrets, such as private keys.
-/// If minor secrets needs to be added, then ensure fields are annotated with `#[serde(skip_serializing)]` to avoid accidentally diclosing them.
-#[derive(Serialize, Deserialize, Clone, Copy, Debug, Default)]
-#[serde(deny_unknown_fields)]
-pub struct CoreToCoreNetworkConfig {
-    pub message_limit: Option<u64>,
-    pub multiplier: Option<f64>,
-    pub max_interval: Option<u64>,
-    pub max_elapsed_time: Option<u64>,
-    /// Initial interval for exponential backoff in milliseconds
-    pub initial_interval_ms: Option<u64>,
-    pub network_timeout: Option<u64>,
-    pub network_timeout_bk: Option<u64>,
-    pub network_timeout_bk_sns: Option<u64>,
-    pub max_en_decode_message_size: Option<u64>,
-    /// Background interval for updating session status
-    pub session_update_interval_secs: Option<u64>,
-    /// Background interval for cleaning up completed sessions
-    pub session_cleanup_interval_secs: Option<u64>,
-    /// Background interval for discarding inactive sessions
-    pub discard_inactive_sessions_interval: Option<u64>,
-    /// Maximum waiting time for trying to push the message in the queue
-    pub max_waiting_time_for_message_queue: Option<u64>,
-    /// Maximum number of "Inactive" sessions a party can open before I refuse to open more
-    pub max_opened_inactive_sessions_per_party: Option<u64>,
-    /// Look-ahead window (in rounds) for buffering future-round messages from a
-    /// peer that is ahead of us: a message more than this many rounds ahead of
-    /// our current round is dropped rather than buffered. Larger tolerates more
-    /// benign reordering/asynchrony but lets a peer make us reserve more memory.
-    /// Should be `>= 1`; `0` disables future buffering and drops legitimate
-    /// reordered messages.
-    pub max_future_rounds: Option<u64>,
-    /// Hard cap on the number of distinct future-round messages buffered per
-    /// sender, independent of `max_future_rounds`. Bounds reordering-buffer
-    /// memory against a peer flooding many distinct future round numbers.
-    /// Should be `>= 1` to tolerate any reordering.
-    pub max_buffered_future_msgs: Option<u64>,
-}
-
-impl CoreToCoreNetworkConfig {
-    pub fn get_message_limit(&self) -> usize {
-        self.message_limit
-            .map(|v| v as usize)
-            .unwrap_or(MESSAGE_LIMIT)
-    }
-
-    pub fn get_multiplier(&self) -> f64 {
-        self.multiplier.unwrap_or(MULTIPLIER)
-    }
-
-    pub fn get_max_interval(&self) -> Duration {
-        self.max_interval
-            .map(Duration::from_secs)
-            .unwrap_or(MAX_INTERVAL)
-    }
-
-    pub fn get_max_elapsed_time(&self) -> Option<Duration> {
-        // Always bounded: an unset value falls back to MAX_ELAPSED_TIME rather
-        // than "retry forever". (The old two-level wrapper returned `None` when
-        // a config was supplied but this field omitted; that partial-config
-        // corner is now defaulted like every other field.)
-        Some(
-            self.max_elapsed_time
-                .map(Duration::from_secs)
-                .unwrap_or(MAX_ELAPSED_TIME),
-        )
-    }
-
-    pub fn get_network_timeout(&self) -> Duration {
-        self.network_timeout
-            .map(Duration::from_secs)
-            .unwrap_or(NETWORK_TIMEOUT_LONG)
-    }
-
-    pub fn get_network_timeout_bk(&self) -> Duration {
-        self.network_timeout_bk
-            .map(Duration::from_secs)
-            .unwrap_or(NETWORK_TIMEOUT_BK)
-    }
-
-    pub fn get_network_timeout_bk_sns(&self) -> Duration {
-        self.network_timeout_bk_sns
-            .map(Duration::from_secs)
-            .unwrap_or(NETWORK_TIMEOUT_BK_SNS)
-    }
-
-    pub fn get_max_en_decode_message_size(&self) -> usize {
-        self.max_en_decode_message_size
-            .map(|v| v as usize)
-            .unwrap_or(*MAX_EN_DECODE_MESSAGE_SIZE)
-    }
-
-    pub fn get_initial_interval(&self) -> Duration {
-        Duration::from_millis(self.initial_interval_ms.unwrap_or(INITIAL_INTERVAL_MS))
-    }
-
-    pub fn get_session_update_interval(&self) -> Duration {
-        Duration::from_secs(
-            self.session_update_interval_secs
-                .unwrap_or(SESSION_STATUS_UPDATE_INTERVAL_SECS),
-        )
-    }
-
-    pub fn get_session_cleanup_interval(&self) -> Duration {
-        Duration::from_secs(
-            self.session_cleanup_interval_secs
-                .unwrap_or(SESSION_CLEANUP_INTERVAL_SECS),
-        )
-    }
-
-    pub fn get_discard_inactive_sessions_interval(&self) -> Duration {
-        Duration::from_secs(
-            self.discard_inactive_sessions_interval
-                .unwrap_or(DISCARD_INACTIVE_SESSION_INTERVAL_SECS),
-        )
-    }
-
-    pub fn get_max_opened_inactive_sessions_per_party(&self) -> u64 {
-        self.max_opened_inactive_sessions_per_party
-            .unwrap_or(MAX_OPENED_INACTIVE_SESSIONS_PER_PARTY)
-    }
-
-    pub fn get_max_waiting_time_for_message_queue(&self) -> Duration {
-        Duration::from_secs(
-            self.max_waiting_time_for_message_queue
-                .unwrap_or(MAX_WAITING_TIME_MESSAGE_QUEUE),
-        )
-    }
-
-    pub fn get_max_buffered_future_msgs(&self) -> usize {
-        self.max_buffered_future_msgs
-            .map(|v| v as usize)
-            .unwrap_or(MAX_BUFFERED_FUTURE_MSGS)
-    }
-}
-
-//TODO: Most likely need this to create NetworkStack instead of GrpcNetworking
-/// GrpcNetworkingManager is responsible for managing
-/// channels and message queues between MPC parties.
-#[derive(Debug, Clone)]
-pub struct GrpcNetworkingManager {
-    // Session reference storage to prevent premature cleanup under high concurrency
-    pub(crate) session_store: Arc<SessionStore>,
-    inactive_session_count: Arc<AtomicU64>,
-    active_session_count: Arc<AtomicU64>,
-    // Keeps tracks of how many sessions were opened by each party
-    // NOTE: Always lock session_store before opened_sessions_tracker to prevent deadlocks
-    pub opened_sessions_tracker: Arc<DashMap<MpcIdentity, u64>>,
-    conf: CoreToCoreNetworkConfig,
-    pub sending_service: GrpcSendingService,
-    #[cfg(feature = "testing")]
-    pub force_tls: bool,
-}
-
-pub type GrpcServer = GnetworkingServer<NetworkingImpl>;
-
-impl GrpcNetworkingManager {
-    /// Create a new server from the networking manager.
-    /// The server can be used as a tower Service.
-    pub fn new_server(
-        &self,
-        tls_extension: TlsExtensionGetter,
-    ) -> GnetworkingServer<NetworkingImpl> {
-        GnetworkingServer::new(NetworkingImpl::new(
-            Arc::clone(&self.session_store),
-            Arc::clone(&self.opened_sessions_tracker),
-            self.conf.get_message_limit(),
-            self.conf.get_max_opened_inactive_sessions_per_party(),
-            self.conf.get_max_waiting_time_for_message_queue(),
-            tls_extension,
-            #[cfg(feature = "testing")]
-            self.force_tls,
-        ))
-        .max_decoding_message_size(self.conf.get_max_en_decode_message_size())
-        .max_encoding_message_size(self.conf.get_max_en_decode_message_size())
-    }
-
-    /// Starts a background task that periodically cleans up the session store, it wakes up at every update_interval.
-    ///
-    /// The task discards sessions that have been completed for longer than the cleanup interval
-    /// and inactive session that have been inactive for longer than the discard_inactive_interval.
-    ///
-    /// It also updates the status of active sessions by checking if their weak references are still valid,
-    /// and if not, marks them as completed.
-    ///
-    /// Finally it also updates the counts of inactive and active sessions.
-    fn start_background_cleaning_task(
-        session_store: Arc<SessionStore>,
-        inactive_session_count: Arc<AtomicU64>,
-        active_session_count: Arc<AtomicU64>,
-        update_interval: Duration,
-        cleanup_interval: Duration,
-        discard_inactive_interval: Duration,
-    ) {
-        tokio::spawn(async move {
-            let mut interval = tokio::time::interval(update_interval);
-            loop {
-                interval.tick().await;
-                let mut internal_inactive_sessions_count = 0;
-                let mut internal_active_sessions_count = 0;
-                let mut to_remove = Vec::new();
-                for mut cur in session_store.iter_mut() {
-                    let (session_id, status) = cur.pair_mut();
-                    match status {
-                        SessionStatus::Completed(started) => {
-                            // Remove completed sessions that have been completed for a very long time
-                            if started.elapsed() > cleanup_interval {
-                                to_remove.push(*session_id);
-                            }
-                        }
-                        SessionStatus::Inactive((_, started)) => {
-                            // Remove inactive sessions that have been inactive for awhile
-                            if started.elapsed() > discard_inactive_interval {
-                                tracing::warn!(
-                                    "Discarding Inactive session {:?} after {:?} seconds. We never heard about such session.",
-                                    session_id,
-                                    started.elapsed().as_secs()
-                                );
-                                to_remove.push(*session_id);
-                                continue;
-                            } else {
-                                internal_inactive_sessions_count += 1;
-                            }
-                        }
-                        SessionStatus::Active(session) => match session.upgrade() {
-                            Some(network_session) => {
-                                let time_since_last_rec = Duration::from_millis(
-                                    now_activity_millis().saturating_sub(
-                                        network_session
-                                            .last_rec_activity_time
-                                            .load(Ordering::Relaxed),
-                                    ),
-                                );
-                                if time_since_last_rec > discard_inactive_interval {
-                                    tracing::warn!(
-                                        "Discarding Active session {:?} after {:?} seconds.",
-                                        session_id,
-                                        time_since_last_rec.as_secs()
-                                    );
-                                    to_remove.push(*session_id);
-                                    continue;
-                                }
-                                internal_active_sessions_count += 1;
-                            }
-                            None => {
-                                *status = SessionStatus::Completed(Instant::now());
-                            }
-                        },
-                    };
-                }
-                for session_id in to_remove {
-                    session_store.remove(&session_id);
-                }
-                inactive_session_count.store(internal_inactive_sessions_count, Ordering::Relaxed);
-                active_session_count.store(internal_active_sessions_count, Ordering::Relaxed);
-            }
-        });
-    }
-
-    /// Owner should be the external address
-    pub fn new(
-        tls_conf: Option<tokio_rustls::rustls::client::ClientConfig>,
-        conf: CoreToCoreNetworkConfig,
-    ) -> anyhow::Result<Self> {
-        #[cfg(feature = "testing")]
-        let force_tls = tls_conf.is_some();
-        #[cfg(feature = "testing")]
-        if !force_tls {
-            tracing::warn!(
-                "force_tls is DISABLED. Testing feature is enabled - this is NOT recommended in production environments."
-            );
-        }
-
-        #[cfg(not(any(test, feature = "testing")))]
-        if tls_conf.is_none() {
-            return Err(error_utils::anyhow_error_and_log(
-                "TLS configuration must be provided in non-testing environments",
-            ));
-        }
-
-        // `conf` is already resolved: an absent config is `CoreToCoreNetworkConfig::default()`
-        // (all fields `None`), so every `get_*` accessor falls back to its constant.
-        let session_store = Arc::new(SessionStore::default());
-
-        // We need to spawn background cleanup task to remove dead weak references from session_store, otherwise they accumulate and eat RAM + perf
-        let cleanup_session_store = Arc::clone(&session_store);
-        let update_interval = conf.get_session_update_interval();
-        let cleanup_interval = conf.get_session_cleanup_interval();
-        let discard_inactive_interval = conf.get_discard_inactive_sessions_interval();
-        let inactive_session_count = Arc::new(AtomicU64::new(0));
-        let active_session_count = Arc::new(AtomicU64::new(0));
-        Self::start_background_cleaning_task(
-            cleanup_session_store,
-            Arc::clone(&inactive_session_count),
-            Arc::clone(&active_session_count),
-            update_interval,
-            cleanup_interval,
-            discard_inactive_interval,
-        );
-
-        Ok(GrpcNetworkingManager {
-            session_store,
-            inactive_session_count,
-            active_session_count,
-            opened_sessions_tracker: Arc::new(DashMap::new()),
-            conf,
-            sending_service: GrpcSendingService::new(tls_conf, conf)?,
-            #[cfg(feature = "testing")]
-            force_tls,
-        })
-    }
-
-    pub async fn make_healthcheck_session<R: RoleTrait>(
-        &self,
-        role_assignment: &RoleAssignment<R>,
-        my_role: R,
-    ) -> anyhow::Result<HealthCheckSession<R>> {
-        let mut others = role_assignment.clone();
-
-        // Removing self from the role_assignment map
-        // as we only want to connect to others.
-        // Store my own identity in the session
-        let owner = match others.remove(&my_role) {
-            Some(owner) => owner,
-            None => {
-                return Err(anyhow::anyhow!(
-                    "My role {:?} not found in role assignment {:?}",
-                    my_role,
-                    role_assignment
-                ));
-            }
-        };
-
-        let mut connection_channels = HashMap::new();
-        for (role, identity) in others.inner.into_iter() {
-            let channel = self.sending_service.connect_to_party(&identity).await?;
-            connection_channels.insert((role, identity), channel);
-        }
-
-        Ok(HealthCheckSession::new(
-            owner,
-            my_role,
-            // We use the same timeout in HealthCheck than
-            // in Sync MPC protocols
-            self.conf.get_network_timeout(),
-            connection_channels,
-        ))
-    }
-
-    /// Create a new session from the network manager.
-    ///
-    /// All the communication are performed using sessions.
-    /// There may be multiple session in parallel,
-    /// identified by different session IDs.
-    pub async fn make_network_session<R: RoleTrait>(
-        &self,
-        session_id: SessionId,
-        role_assignment: &RoleAssignment<R>,
-        my_role: R,
-        network_mode: NetworkMode,
-    ) -> anyhow::Result<Arc<impl Networking<R> + use<R>>> {
-        let party_count = role_assignment.len();
-        let mut others = role_assignment.clone();
-
-        // Removing self from the role_assignment map
-        // as we only want to connect to others.
-        // Store my own identity in the session
-        let owner = match others.remove(&my_role) {
-            Some(owner) => owner,
-            None => {
-                return Err(anyhow::anyhow!(
-                    "My role {:?} not found in role assignment {:?}",
-                    my_role,
-                    role_assignment
-                ));
-            }
-        };
-
-        let timeout = match network_mode {
-            // If we have waited for the network as long as we keep inactive sessions, we might as well timeout
-            NetworkMode::Async => self.conf.get_discard_inactive_sessions_interval(),
-            NetworkMode::Sync => self.conf.get_network_timeout(),
-        };
-
-        let (connection_channel, completed_parties) =
-            self.sending_service.add_connections(&others).await?;
-
-        let session = match self.session_store.entry(session_id) {
-            // Turn an inactive session into an active one
-            dashmap::Entry::Occupied(mut status) => {
-                let mutable_status = status.get_mut();
-
-                let message_store = if let SessionStatus::Inactive(message_store) = mutable_status {
-                    // Upgrade the message store from the uninitialized state to the initialized state
-                    message_store.0.init(
-                        self.conf.get_message_limit(),
-                        &others,
-                        Arc::clone(&self.opened_sessions_tracker),
-                    );
-                    message_store.clone()
-                } else {
-                    return Err(anyhow::anyhow!(
-                        "Session {:?} already exists and is not inactive for {}",
-                        session_id,
-                        owner
-                    ));
-                };
-
-                let session = Arc::new(NetworkSession::new(
-                    owner.clone(),
-                    session_id,
-                    connection_channel,
-                    message_store.0,
-                    completed_parties,
-                    network_mode,
-                    self.conf,
-                    timeout,
-                ));
-
-                *mutable_status = SessionStatus::Active(Arc::downgrade(&session));
-
-                session
-            }
-            dashmap::Entry::Vacant(vacant) => {
-                let message_queue = MessageQueueStore::new_initialized(
-                    self.conf.get_message_limit(),
-                    &others,
-                    Arc::clone(&self.opened_sessions_tracker),
-                );
-
-                let session = Arc::new(NetworkSession::new(
-                    owner.clone(),
-                    session_id,
-                    connection_channel,
-                    message_queue,
-                    completed_parties,
-                    network_mode,
-                    self.conf,
-                    timeout,
-                ));
-
-                vacant.insert(SessionStatus::Active(Arc::downgrade(&session)));
-
-                session
-            }
-        };
-
-        tracing::info!(
-            "[SESSION_CREATION] Starting session {:?} with {} parties. (Owner: {:?})",
-            session_id,
-            party_count,
-            owner,
-        );
-
-        Ok(session)
-    }
-
-    pub async fn active_session_count(&self) -> u64 {
-        self.active_session_count.load(Ordering::Relaxed)
-    }
-
-    pub async fn inactive_session_count(&self) -> u64 {
-        self.inactive_session_count.load(Ordering::Relaxed)
-    }
-}
-
-// we need a counter for each value sent over the local queues
-// so that messages that haven't been pickup up using receive() calls will get dropped
-#[derive(Debug)]
-pub struct NetworkRoundValue {
-    pub value: Vec<u8>,
-    pub round_counter: usize,
-}
-
-/// Per-sender receiver plus a bounded buffer of future-round messages that
-/// arrived before the local session advanced to their round.
-///
-/// The receiver is a bounded mpsc whose ordering is fully controlled by the
-/// (possibly malicious) peer, so `receive` must not treat "next packet" as
-/// "current-round packet". Messages tagged with a strictly future round are
-/// parked here — keyed by their round counter — until the session reaches that
-/// round, at which point they are delivered from the fast path. The buffer is
-/// bounded by [`MAX_BUFFERED_FUTURE_MSGS`] to prevent
-/// a peer from exhausting memory with distinct future round numbers.
-#[derive(Debug)]
-pub(crate) struct ReceiverState {
-    /// The bounded mpsc receiver for this sender.
-    pub(crate) rx: Receiver<NetworkRoundValue>,
-    /// Future-round messages, keyed by round counter (all strictly greater than
-    /// the current round when inserted). Holds at most one value per round.
-    pub(crate) future: std::collections::BTreeMap<usize, Vec<u8>>,
-}
-
-impl ReceiverState {
-    /// Wrap a freshly created receiver with an empty future-round buffer.
-    pub(crate) fn new(rx: Receiver<NetworkRoundValue>) -> Self {
-        Self {
-            rx,
-            future: std::collections::BTreeMap::new(),
-        }
-    }
-
-    /// Drop every buffered message strictly older than `current`, then take the
-    /// buffered message for exactly `current` if one is present.
-    pub(crate) fn take_current(&mut self, current: usize) -> Option<Vec<u8>> {
-        // Keep only rounds >= current; everything below is stale and dropped.
-        self.future = self.future.split_off(&current);
-        self.future.remove(&current)
-    }
-
-    /// Buffer a strictly-future-round message, enforcing the reordering-buffer
-    /// bounds. Returns `true` if the message was buffered, `false` if it was
-    /// dropped for being outside the `max_future_rounds` look-ahead window or
-    /// over the per-sender `max_buffered` cap.
-    ///
-    /// The two bounds are passed in (resolved from
-    /// [`CoreToCoreNetworkConfig`](crate::grpc::CoreToCoreNetworkConfig)) rather
-    /// than read from constants, so buffering tolerance is configurable while
-    /// this method stays free of any config dependency and unit-testable in
-    /// isolation. Their defaults are [`MAX_FUTURE_ROUNDS`] / [`MAX_BUFFERED_FUTURE_MSGS`].
-    ///
-    /// At most one value is kept per round: a duplicate for an already-buffered
-    /// round is discarded (first one wins) so a malicious peer cannot clobber a
-    /// genuine future-round message.
-    pub(crate) fn buffer_future(
-        &mut self,
-        round: usize,
-        value: Vec<u8>,
-        current: usize,
-        max_buffered: usize,
-    ) -> bool {
-        if round <= current.saturating_add(max_buffered) && self.future.len() < max_buffered {
-            self.future.entry(round).or_insert(value);
-            true
-        } else {
-            false
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct InitializedMessageQueueStore {
-    // role assignment is needed because the message store
-    // needs to translate between identity and role
-    tx: DashMap<MpcIdentity, Arc<Sender<NetworkRoundValue>>>,
-    receiver_state: DashMap<RoleKind, Arc<Mutex<ReceiverState>>>,
-}
-
-/// The pair of channel halves kept per sender in an uninitialized message
-/// queue: the sending half other parties push into (`tx`) and the shared
-/// receiving state the owning session drains (`rx`).
-#[derive(Debug, Clone)]
-pub(crate) struct ChannelPair {
-    pub(crate) tx: Arc<Sender<NetworkRoundValue>>,
-    pub(crate) rx: Arc<Mutex<ReceiverState>>,
-}
-
-#[derive(Debug, Clone)]
-pub(crate) enum MessageQueueStore {
-    Uninitialized(DashMap<MpcIdentity, ChannelPair>),
-    Initialized(InitializedMessageQueueStore),
-}
-
-impl MessageQueueStore {
-    pub(crate) fn new_uninitialized(channel_maps: DashMap<MpcIdentity, ChannelPair>) -> Self {
-        MessageQueueStore::Uninitialized(channel_maps)
-    }
-
-    pub(crate) fn new_initialized<R: RoleTrait>(
-        channel_size_limit: usize,
-        others: &RoleAssignment<R>,
-        opened_sessions_tracker: Arc<DashMap<MpcIdentity, u64>>,
-    ) -> Self {
-        let mut out = Self::new_uninitialized(DashMap::new());
-        out.init(channel_size_limit, others, opened_sessions_tracker);
-        out
-    }
-
-    pub(crate) fn init<R: RoleTrait>(
-        &mut self,
-        channel_size_limit: usize,
-        others: &RoleAssignment<R>,
-        opened_sessions_tracker: Arc<DashMap<MpcIdentity, u64>>,
-    ) {
-        if let MessageQueueStore::Uninitialized(channel_maps) = &self {
-            let tx_map = DashMap::with_capacity(channel_maps.len());
-            let rx_map = DashMap::with_capacity(channel_maps.len());
-
-            // If an identity is in channel_maps and also in role_assignment,
-            // then we insert it to the initialized message queue.
-            // If an identity is not in channel_maps but in role_assignment,
-            // then we create a new channel for it.
-            for (role, identity) in others.iter() {
-                let mpc_id = identity.mpc_identity();
-                if let Some(entry) = channel_maps.get(&mpc_id) {
-                    opened_sessions_tracker
-                        .entry(mpc_id.clone())
-                        .and_modify(|count| {
-                            *count = count.saturating_sub(1);
-                        })
-                        .or_insert(0);
-                    let pair = entry.value();
-                    tx_map.insert(entry.key().clone(), pair.tx.clone());
-                    rx_map.insert(role.get_role_kind(), pair.rx.clone());
-                } else {
-                    let (tx, rx) = channel::<NetworkRoundValue>(channel_size_limit);
-                    tx_map.insert(identity.mpc_identity(), Arc::new(tx));
-                    rx_map.insert(
-                        role.get_role_kind(),
-                        Arc::new(Mutex::new(ReceiverState::new(rx))),
-                    );
-                }
-            }
-
-            *self = MessageQueueStore::Initialized(InitializedMessageQueueStore {
-                tx: tx_map,
-                receiver_state: rx_map,
-            });
-        } else {
-            tracing::warn!("MessageQueueStore is already initialized");
-        }
-    }
-
-    pub(crate) fn get_tx(
-        &self,
-        mpc_identity: &MpcIdentity,
-    ) -> Result<Option<Arc<Sender<NetworkRoundValue>>>, Box<tonic::Status>> {
-        match &self {
-            MessageQueueStore::Initialized(store) => Ok(store
-                .tx
-                .get(mpc_identity)
-                .map(|entry| entry.value().clone())),
-            MessageQueueStore::Uninitialized(_) => Err(Box::new(tonic::Status::internal(
-                "trying to get tx message queue on id {mpc_identity} when it is not initialized",
-            ))),
-        }
-    }
-
-    pub(crate) fn get_receiver_state<R: RoleTrait>(
-        &self,
-        role: &R,
-    ) -> anyhow::Result<Option<Arc<Mutex<ReceiverState>>>> {
-        match &self {
-            MessageQueueStore::Initialized(store) => Ok(store
-                .receiver_state
-                .get(&role.get_role_kind())
-                .map(|entry| entry.value().clone())),
-            MessageQueueStore::Uninitialized(_) => Err(anyhow::anyhow!(
-                "trying to get rx message queue on role {role} when it is not initialized",
-            )),
-        }
-    }
-
-    // this must be performed on the uninitialized message queue
-    pub(crate) fn entry(
-        &self,
-        mpc_identity: MpcIdentity,
-    ) -> anyhow::Result<dashmap::mapref::entry::Entry<'_, MpcIdentity, ChannelPair>> {
-        match &self {
-            MessageQueueStore::Uninitialized(inner) => Ok(inner.entry(mpc_identity)),
-            MessageQueueStore::Initialized(_) => Err(anyhow::anyhow!(
-                "entry can only be performed on uninitialized message queue"
-            )),
-        }
-    }
-
-    pub(crate) fn iter_keys(
-        &self,
-    ) -> Result<impl Iterator<Item = RoleKind> + '_, Box<tonic::Status>> {
-        match &self {
-            MessageQueueStore::Uninitialized(_) => Err(Box::new(tonic::Status::internal(
-                "trying to iterate keys when message queue is not initialized",
-            ))),
-            MessageQueueStore::Initialized(inner) => {
-                Ok(inner.receiver_state.iter().map(|entry| *entry.key()))
-            }
-        }
-    }
-}
-
-pub(crate) type SessionStore = DashMap<SessionId, SessionStatus>;
-
-#[derive(Debug)]
-/// Represents the status of a session in the session store.
-/// It can be:
-/// - Completed: The session has been completed and the timestamp of completion is stored.
-/// - Inactive: The session is inactive (I haven't yet heard about the request) and has a message queue store for senders.
-/// - Active: The session is active (I know about the request) and holds a weak reference to the `NetworkSession`.
-pub(crate) enum SessionStatus {
-    Completed(Instant),
-    Inactive((MessageQueueStore, Instant)),
-    Active(Weak<NetworkSession>),
-}
-
-impl From<SessionStatus> for SendValueResponse {
-    fn from(value: SessionStatus) -> Self {
-        // Same mapping as the borrowing impl; delegate so the two cannot drift.
-        (&value).into()
-    }
-}
-
-impl From<&SessionStatus> for SendValueResponse {
-    fn from(value: &SessionStatus) -> Self {
-        let status = match value {
-            SessionStatus::Completed(_) => Status::Completed,
-            SessionStatus::Inactive(_) => Status::Inactive,
-            SessionStatus::Active(_) => Status::Active,
-        };
-        SendValueResponse {
-            status: status.into(),
-        }
-    }
-}
-
-// Because we can use a custom TCP Incoming, we need to specify how
-// to extract the TLS extension from the incoming connection
-#[derive(Default)]
-pub enum TlsExtensionGetter {
-    #[default]
-    TlsConnectInfo,
-    SslConnectInfo,
-}
-
-#[derive(Default)]
-pub struct NetworkingImpl {
-    session_store: Arc<SessionStore>,
-    // key is the MPC identity
-    opened_sessions_tracker: Arc<DashMap<MpcIdentity, u64>>,
-    channel_size_limit: usize,
-    max_opened_inactive_sessions: u64,
-    max_waiting_time_for_message_queue: Duration,
-    tls_extension: TlsExtensionGetter,
-    // We gate this behind the testing feature because in non-testing environments
-    // we want to ALWAYS use TLS for security reasons.
-    #[cfg(feature = "testing")]
-    force_tls: bool,
-}
-
-impl NetworkingImpl {
-    fn new(
-        session_store: Arc<SessionStore>,
-        opened_sessions_tracker: Arc<DashMap<MpcIdentity, u64>>,
-        channel_size_limit: usize,
-        max_opened_inactive_sessions: u64,
-        max_waiting_time_for_message_queue: Duration,
-        tls_extension: TlsExtensionGetter,
-        #[cfg(feature = "testing")] force_tls: bool,
-    ) -> Self {
-        Self {
-            session_store: session_store.clone(),
-            opened_sessions_tracker: opened_sessions_tracker.clone(),
-            channel_size_limit,
-            max_opened_inactive_sessions,
-            max_waiting_time_for_message_queue,
-            tls_extension,
-            #[cfg(feature = "testing")]
-            force_tls,
-        }
-    }
-
-    /// Extract and parse the sender identity from the request's TLS certificate,
-    /// according to the configured [`TlsExtensionGetter`].
-    ///
-    /// Returns `Ok(None)` when no client certificate is present (whether that is
-    /// acceptable is decided later by [`sender_verification`]). This is the single
-    /// source of truth for identity extraction so `health_check` and `send_value`
-    /// — which must perform the *exact same* check — cannot diverge.
-    fn extract_tls_sender<T>(
-        &self,
-        request: &tonic::Request<T>,
-    ) -> Result<Option<String>, tonic::Status> {
-        match self.tls_extension {
-            TlsExtensionGetter::TlsConnectInfo => request
-                .extensions()
-                .get::<tonic::transport::server::TlsConnectInfo<TcpConnectInfo>>()
-                .and_then(|i| i.peer_certs().map(parse_identity_from_cert)),
-            TlsExtensionGetter::SslConnectInfo => request
-                .extensions()
-                .get::<tonic_tls::rustls::SslConnectInfo<TcpConnectInfo>>()
-                .and_then(|i| i.peer_certs().map(parse_identity_from_cert)),
-        }
-        .transpose()
-        .map_err(|boxed| *boxed)
-    }
-
-    /// Fetches the channel for the given session and tag.
-    /// - If the session is inactive, it creates a new channel for the sender (assuming the sender hasn't opened too many channels for inactive sessions yet).
-    /// - If the session is active, it returns the existing channel (assuming the sender is part of the session).
-    /// - If the session is completed, it returns None
-    ///   to indicate that the message can be accepted but will not be processed.
-    fn fetch_tx_channel(
-        &self,
-        session_status: &SessionStatus,
-        tag: &Tag,
-    ) -> Result<Option<Arc<Sender<NetworkRoundValue>>>, tonic::Status> {
-        match session_status {
-            SessionStatus::Completed(_) => {
-                tracing::debug!(
-                    "Session {:?} found in session_store but is completed. Will be removed by background cleanup.",
-                    tag.session_id
-                );
-                // We accept the message even if we won't do anything with it
-                // to avoid blocking the sender
-                Ok(None)
-            }
-            // Session is inactive, we may need to create a new channel for the sender
-            SessionStatus::Inactive(message_queue) => {
-                tracing::debug!(
-                    "Session {:?} found in session_store but is inactive.",
-                    tag.session_id
-                );
-                // Check if the sender already has a channel, if it does, return it
-                // if it doesn't then create a new one.
-                // Note that we need to do this atomically to avoid race conditions.
-                match message_queue.0.entry(tag.sender.clone()).map_err(|e| {
-                    tonic::Status::internal(format!(
-                        "Failed to access message queue for session {:?}: {}",
-                        tag.session_id, e
-                    ))
-                })? {
-                    dashmap::Entry::Occupied(occupied_entry) => {
-                        Ok(Some(occupied_entry.get().tx.clone()))
-                    }
-                    dashmap::Entry::Vacant(vacant_entry_tx) => {
-                        let mut opened_session_tracker_entry = self
-                            .opened_sessions_tracker
-                            .entry(tag.sender.clone())
-                            .or_insert(0);
-                        if *opened_session_tracker_entry >= self.max_opened_inactive_sessions {
-                            tracing::warn!(
-                                "Too many inactive sessions opened by {:?}. Have {}, Max allowed: {}",
-                                tag.sender,
-                                *opened_session_tracker_entry,
-                                self.max_opened_inactive_sessions
-                            );
-                            return Err(tonic::Status::new(
-                                tonic::Code::ResourceExhausted,
-                                format!(
-                                    "Too many inactive sessions opened by {:?}. Have {}, Max allowed: {}",
-                                    tag.sender,
-                                    *opened_session_tracker_entry,
-                                    self.max_opened_inactive_sessions
-                                ),
-                            ));
-                        }
-                        // Create a new channel for the sender
-                        let (tx, rx) = channel::<NetworkRoundValue>(self.channel_size_limit);
-                        let tx = Arc::new(tx);
-                        vacant_entry_tx.insert(ChannelPair {
-                            tx: Arc::clone(&tx),
-                            rx: Arc::new(Mutex::new(ReceiverState::new(rx))),
-                        });
-
-                        // Update the opened sessions tracker
-                        *opened_session_tracker_entry += 1;
-                        Ok(Some(tx))
-                    }
-                }
-            }
-            // Session is active, we can proceed with sending the message
-            SessionStatus::Active(weak_session) => {
-                tracing::debug!(
-                    "Session {:?} found in session_store and is active.",
-                    tag.session_id
-                );
-                // Attempt to upgrade weak reference to strong reference
-                if let Some(session) = weak_session.upgrade() {
-                    // Get the message queue from the session's receiving channels
-                    if let Some(session_store) = session
-                        .receiving_channels
-                        .get_tx(&tag.sender)
-                        .map_err(|e| *e)?
-                    {
-                        Ok(Some(session_store.clone()))
-                    } else {
-                        let available_senders: Vec<_> = session
-                            .receiving_channels
-                            .iter_keys()
-                            .map_err(|e| *e)?
-                            .collect();
-
-                        tracing::warn!(
-                            "Sender {:?} not found in session {:?}. Available senders: {:?}",
-                            tag.sender,
-                            tag.session_id,
-                            available_senders
-                        );
-
-                        Err(tonic::Status::new(
-                            tonic::Code::NotFound,
-                            format!(
-                                "Sender {:?} not found in session {:?}",
-                                tag.sender, tag.session_id
-                            ),
-                        ))
-                    }
-                } else {
-                    // Session has been dropped, accept the message even if we won't do anything with it
-                    Ok(None)
-                }
-            }
-        }
-    }
-}
-
-// We do the measurement of received bytes here because
-// some messages may never reach the application level
-// (i.e. in the Networking trait)
-pub static NETWORK_RECEIVED_MEASUREMENT: LazyLock<DashMap<SessionId, usize>> =
-    LazyLock::new(DashMap::new);
-
-fn parse_identity_from_cert(
-    certs: Arc<Vec<CertificateDer<'static>>>,
-) -> Result<String, Box<tonic::Status>> {
-    if certs.len() != 1 {
-        // it shouldn't happen because we expect TLS certificates to
-        // be signed by party CA certificates directly, without any
-        // intermediate CAs
-        tracing::warn!("Received more than one certificate from peer, checking the first one only");
-    }
-
-    parse_x509_certificate(certs[0].as_ref())
-        .map_err(|e| Box::new(tonic::Status::new(tonic::Code::Aborted, e.to_string())))
-        .and_then(|(_rem, cert)| {
-            extract_subject_from_cert(&cert)
-                .map_err(|e| Box::new(tonic::Status::new(tonic::Code::Aborted, e.to_string())))
-        })
-}
-
-// Verify that the sender in the tag matches the identity extracted from the TLS certificate
-fn sender_verification(
-    #[cfg(feature = "testing")] force_tls: bool,
-    tag_sender: &MpcIdentity,
-    valid_tls_sender: Option<String>,
-) -> Result<(), Box<tonic::Status>> {
-    if let Some(sender) = valid_tls_sender {
-        // tag.sender is an the MPC identity, this should match the CN in the certificate
-        if sender != tag_sender.0 {
-            return Err(Box::new(tonic::Status::new(
-                tonic::Code::Unauthenticated,
-                format!(
-                    "wrong sender: expected {sender} to be in in tag {}",
-                    tag_sender
-                ),
-            )));
-        }
-        tracing::debug!("TLS Check went fine for sender: {:?}", sender);
-    } else {
-        // With testing feature, TLS is optional
-        #[cfg(feature = "testing")]
-        {
-            if force_tls {
-                // If force_tls is enabled, we require a TLS certificate
-                tracing::error!("Force TLS is enabled, but no certificate found in the request.");
-                return Err(Box::new(tonic::Status::new(
-                    tonic::Code::Unauthenticated,
-                    "Could not find a TLS certificate in the request to verify user's identity."
-                        .to_string(),
-                )));
-            } else {
-                // since we log this on _every_ send call and only use this for testing builds, we use debug level to reduce log spam
-                tracing::debug!("Force TLS is disabled, and no certificate found in the request.");
-            }
-        }
-
-        // Without testing feature, TLS is mandatory
-        #[cfg(not(any(test, feature = "testing")))]
-        {
-            tracing::error!(
-                "Could not find a TLS certificate in the request to verify user's identity."
-            );
-            return Err(Box::new(tonic::Status::new(
-                tonic::Code::Unauthenticated,
-                "Could not find a TLS certificate in the request to verify user's identity."
-                    .to_string(),
-            )));
-        }
-    }
-    Ok(())
-}
-
-#[async_trait]
-impl Gnetworking for NetworkingImpl {
-    async fn health_check(
-        &self,
-        request: tonic::Request<HealthCheckRequest>,
-    ) -> std::result::Result<tonic::Response<HealthCheckResponse>, tonic::Status> {
-        // Perform the exact same check as we do for a "real" MPC message
-        let valid_tls_sender = self.extract_tls_sender(&request)?;
-        let request = request.into_inner();
-        let health_tag = bc2wrap::deserialize_slice::<HealthTag>(&request.tag).map_err(|e| {
-            tonic::Status::new(
-                tonic::Code::Aborted,
-                format!("failed to parse value: {} as a HealthTag", e),
-            )
-        })?;
-
-        sender_verification(
-            #[cfg(feature = "testing")]
-            self.force_tls,
-            &health_tag.sender,
-            valid_tls_sender,
-        )
-        .map_err(|e| *e)?;
-
-        tracing::debug!("Received a HealthPing from {}", health_tag.sender);
-        Ok(tonic::Response::new(HealthCheckResponse::default()))
-    }
-
-    async fn send_value(
-        &self,
-        request: tonic::Request<SendValueRequest>,
-    ) -> std::result::Result<tonic::Response<SendValueResponse>, tonic::Status> {
-        // If TLS is enabled, A SAN may look like:
-        // DNS:party1.com, IP Address:127.0.0.1, DNS:localhost, IP Address:192.168.0.1, IP Address:0:0:0:0:0:0:0:1
-        // which is a collection of DNS names and IP addresses.
-        // The DNS component must match the "tag" that's in the request for identity verification,
-        // in this case it's party1.com.
-        // We also require party1.com to be the subject and the issuer CN too,
-        // since we're using self-signed certificates at the moment.
-        let valid_tls_sender = self.extract_tls_sender(&request)?;
-
-        let request = request.into_inner();
-        let tag = bc2wrap::deserialize_slice::<Tag>(&request.tag).map_err(|e| {
-            tonic::Status::new(
-                tonic::Code::Aborted,
-                format!("failed to parse value: {}", e),
-            )
-        })?;
-
-        sender_verification(
-            #[cfg(feature = "testing")]
-            self.force_tls,
-            &tag.sender,
-            valid_tls_sender,
-        )
-        .map_err(|e| *e)?;
-
-        tracing::debug!(
-            "Starting session lookup for session_id={:?}, sender={:?}, round={}",
-            tag.session_id,
-            tag.sender,
-            tag.round_counter
-        );
-
-        match NETWORK_RECEIVED_MEASUREMENT.entry(tag.session_id) {
-            dashmap::Entry::Occupied(mut occupied_entry) => {
-                let entry = occupied_entry.get_mut();
-                *entry += request.tag.len() + request.value.len()
-            }
-            dashmap::Entry::Vacant(vacant_entry) => {
-                vacant_entry.insert(request.tag.len() + request.value.len());
-            }
-        };
-
-        // First try with only read lock to avoid blocking
-        let tx = if let Some(session_status) = self.session_store.get(&tag.session_id) {
-            let status = session_status.value();
-            match self.fetch_tx_channel(status, &tag)? {
-                Some(tx) => tx,
-                None => {
-                    // If the session is completed or inactive, we return early
-                    return Ok(tonic::Response::new(status.into()));
-                }
-            }
-        } else {
-            // We write lock the session store to create a new one
-            match self.session_store.entry(tag.session_id) {
-                dashmap::Entry::Occupied(occupied_entry) => {
-                    // Can be occupied if ever state has changed by the time we reach this branch of the if statement
-                    let status = occupied_entry.get();
-                    match self.fetch_tx_channel(status, &tag)? {
-                        Some(tx) => tx,
-                        None => {
-                            // If the session is completed or inactive, we return early
-                            return Ok(tonic::Response::new(status.into()));
-                        }
-                    }
-                }
-                dashmap::Entry::Vacant(vacant_entry) => {
-                    tracing::debug!(
-                        "Session {:?} not found in session_store, creating a new inactive one.",
-                        tag.session_id
-                    );
-                    let mut opened_session_tracker_entry = self
-                        .opened_sessions_tracker
-                        .entry(tag.sender.clone())
-                        .or_insert(0);
-                    if *opened_session_tracker_entry >= self.max_opened_inactive_sessions {
-                        tracing::warn!(
-                            "Too many inactive sessions opened by {:?}. Got {}, Max allowed: {}",
-                            &tag.sender,
-                            *opened_session_tracker_entry,
-                            self.max_opened_inactive_sessions
-                        );
-                        return Err(tonic::Status::new(
-                            tonic::Code::ResourceExhausted,
-                            format!(
-                                "Too many inactive sessions opened by {:?}. Got {}, Max allowed: {}",
-                                tag.sender,
-                                *opened_session_tracker_entry,
-                                self.max_opened_inactive_sessions
-                            ),
-                        ));
-                    }
-                    // Create a new session with an inactive status
-                    let channel_maps = DashMap::new();
-                    let (tx, rx) = channel::<NetworkRoundValue>(self.channel_size_limit);
-                    let tx = Arc::new(tx);
-                    channel_maps.insert(
-                        tag.sender.clone(),
-                        ChannelPair {
-                            tx: Arc::clone(&tx),
-                            rx: Arc::new(Mutex::new(ReceiverState::new(rx))),
-                        },
-                    );
-
-                    // Insert the new session into the store
-                    vacant_entry.insert(SessionStatus::Inactive((
-                        MessageQueueStore::new_uninitialized(channel_maps),
-                        Instant::now(),
-                    )));
-                    *opened_session_tracker_entry += 1;
-                    tx
-                }
-            }
-        };
-
-        // Send message - ignore send errors as receiver may have dropped
-        let send_result = tokio::time::timeout(
-            self.max_waiting_time_for_message_queue,
-            tx.send(NetworkRoundValue {
-                value: request.value,
-                // Narrow the fixed-width wire round back to the in-memory `usize`.
-                round_counter: tag.round_counter as usize,
-            }),
-        )
-        .await;
-
-        if let Err(e) = send_result {
-            tracing::warn!(
-                "Failed to process value for session {:?}, sender {:?}, round {}. Queue has been full for {} seconds.",
-                tag.session_id,
-                &tag.sender,
-                tag.round_counter,
-                self.max_waiting_time_for_message_queue.as_secs()
-            );
-
-            return Err(tonic::Status::new(
-                tonic::Code::ResourceExhausted,
-                format!(
-                    "Failed to process value for session {:?}, sender {:?}, round {}: {:?}",
-                    tag.session_id, tag.sender, tag.round_counter, e
-                ),
-            ));
-        }
-
-        Ok(tonic::Response::new(SendValueResponse {
-            status: Status::Active.into(),
-        }))
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct Tag {
-    pub(crate) session_id: SessionId,
-    pub(crate) sender: MpcIdentity,
-    /// Round the message belongs to. This is a *wire* field, so it is a
-    /// fixed-width `u64` rather than a platform-dependent `usize`: the encoding
-    /// (bincode legacy/fixint via `bc2wrap`) must not vary with the sender's
-    /// pointer width. The in-memory round counter is `usize`; conversions happen
-    /// at the send/receive boundary.
-    pub(crate) round_counter: u64,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct HealthTag {
-    pub(crate) sender: MpcIdentity,
-}
-
-#[cfg(test)]
-mod tests {
-    use std::sync::Arc;
-
-    use super::*;
-    use dashmap::DashMap;
-    use threshold_types::party::{Identity, RoleAssignment};
-    use threshold_types::role::Role;
-    use tokio::sync::Mutex;
-    use tokio::sync::mpsc::channel;
-    use tokio::time::{Duration, Instant};
-
-    #[test]
-    fn test_fetch_tx_channel_completed_session() {
-        let session_store: Arc<SessionStore> = Arc::new(DashMap::new());
-        let opened_sessions_tracker: Arc<DashMap<MpcIdentity, u64>> = Arc::new(DashMap::new());
-        let session_id = SessionId::from(1u128);
-
-        // Insert a completed session
-        session_store.insert(session_id, SessionStatus::Completed(Instant::now()));
-
-        let networking_impl = NetworkingImpl::new(
-            session_store.clone(),
-            opened_sessions_tracker,
-            100,
-            50,
-            Duration::from_secs(60),
-            TlsExtensionGetter::default(),
-            #[cfg(feature = "testing")]
-            false,
-        );
-
-        let tag = Tag {
-            session_id,
-            sender: MpcIdentity("party1".to_string()),
-            round_counter: 0,
-        };
-        let session_status = session_store.get(&session_id).unwrap();
-
-        let result = networking_impl.fetch_tx_channel(session_status.value(), &tag);
-
-        // Should return Ok(None) for completed sessions
-        assert!(result.is_ok());
-        assert!(result.unwrap().is_none());
-    }
-
-    #[test]
-    fn test_inactive_session_creates_new_channel() {
-        let session_store: Arc<SessionStore> = Arc::new(DashMap::new());
-        let opened_sessions_tracker: Arc<DashMap<MpcIdentity, u64>> = Arc::new(DashMap::new());
-        let session_id = SessionId::from(2u128);
-
-        // Insert an inactive session with empty message queue
-        let channel_maps = DashMap::new();
-        session_store.insert(
-            session_id,
-            SessionStatus::Inactive((
-                MessageQueueStore::new_uninitialized(channel_maps),
-                Instant::now(),
-            )),
-        );
-
-        let networking_impl = NetworkingImpl::new(
-            session_store.clone(),
-            opened_sessions_tracker.clone(),
-            100,
-            50,
-            Duration::from_secs(60),
-            TlsExtensionGetter::default(),
-            #[cfg(feature = "testing")]
-            false,
-        );
-
-        let tag = Tag {
-            session_id,
-            sender: MpcIdentity("party1".to_string()), // Sender does not exist
-            round_counter: 0,
-        };
-        let session_status = session_store.get(&session_id).unwrap();
-
-        let result = networking_impl.fetch_tx_channel(session_status.value(), &tag);
-
-        // Should return Ok(Some(tx)) for inactive sessions with new sender
-        assert!(result.is_ok());
-        assert!(result.unwrap().is_some());
-
-        // Check that opened_sessions_tracker was updated
-        assert_eq!(
-            *opened_sessions_tracker.get(&tag.sender).unwrap().value(),
-            1
-        );
-    }
-
-    #[test]
-    fn test_inactive_session_existing_sender() {
-        let session_store: Arc<SessionStore> = Arc::new(DashMap::new());
-        let opened_sessions_tracker: Arc<DashMap<MpcIdentity, u64>> = Arc::new(DashMap::new());
-        let session_id = SessionId::from(3u128);
-        let sender = MpcIdentity("party1".to_string());
-
-        // Insert an inactive session with existing sender channel
-        let channel_maps = DashMap::new();
-        let (tx, rx) = channel::<NetworkRoundValue>(100);
-        channel_maps.insert(
-            sender.clone(),
-            ChannelPair {
-                tx: Arc::new(tx),
-                rx: Arc::new(Mutex::new(ReceiverState::new(rx))),
-            },
-        );
-
-        session_store.insert(
-            session_id,
-            SessionStatus::Inactive((
-                MessageQueueStore::new_uninitialized(channel_maps),
-                Instant::now(),
-            )),
-        );
-
-        let networking_impl = NetworkingImpl::new(
-            session_store.clone(),
-            opened_sessions_tracker.clone(),
-            100,
-            50,
-            Duration::from_secs(60),
-            TlsExtensionGetter::default(),
-            #[cfg(feature = "testing")]
-            false,
-        );
-
-        let tag = Tag {
-            session_id,
-            sender: MpcIdentity("party1".to_string()),
-            round_counter: 0,
-        };
-        let session_status = session_store.get(&session_id).unwrap();
-
-        let result = networking_impl.fetch_tx_channel(session_status.value(), &tag);
-
-        // Should return Ok(Some(tx)) for inactive sessions with existing sender
-        assert!(result.is_ok());
-        assert!(result.unwrap().is_some());
-
-        // opened_sessions_tracker should NOT be updated since sender already existed
-        assert!(opened_sessions_tracker.get(&tag.sender).is_none());
-    }
-
-    #[test]
-    fn test_too_many_inactive_sessions() {
-        let session_store: Arc<SessionStore> = Arc::new(DashMap::new());
-        let opened_sessions_tracker: Arc<DashMap<MpcIdentity, u64>> = Arc::new(DashMap::new());
-        let session_id = SessionId::from(4u128);
-        let sender = MpcIdentity("party1".to_string());
-
-        // Pre-fill the tracker to hit the limit
-        opened_sessions_tracker.insert(sender.clone(), 50);
-
-        // Insert an inactive session with empty message queue
-        let channel_maps = DashMap::new();
-        session_store.insert(
-            session_id,
-            SessionStatus::Inactive((
-                MessageQueueStore::new_uninitialized(channel_maps),
-                Instant::now(),
-            )),
-        );
-
-        let networking_impl = NetworkingImpl::new(
-            session_store.clone(),
-            opened_sessions_tracker.clone(),
-            100,
-            50, // max_opened_inactive_sessions
-            Duration::from_secs(60),
-            TlsExtensionGetter::default(),
-            #[cfg(feature = "testing")]
-            false,
-        );
-
-        let tag = Tag {
-            session_id,
-            sender: MpcIdentity("party1".to_string()),
-            round_counter: 0,
-        };
-        let session_status = session_store.get(&session_id).unwrap();
-
-        let result = networking_impl.fetch_tx_channel(session_status.value(), &tag);
-
-        // Should return Err with ResourceExhausted
-        assert!(result.is_err());
-        let err = result.unwrap_err();
-        assert_eq!(err.code(), tonic::Code::ResourceExhausted);
-    }
-
-    #[test]
-    fn test_active_session_weak_reference_dead() {
-        let session_store: Arc<SessionStore> = Arc::new(DashMap::new());
-        let opened_sessions_tracker: Arc<DashMap<MpcIdentity, u64>> = Arc::new(DashMap::new());
-        let session_id = SessionId::from(5u128);
-
-        // Create a weak reference that's already dead (no strong reference)
-        let weak: Weak<NetworkSession> = Weak::new();
-        session_store.insert(session_id, SessionStatus::Active(weak));
-
-        let networking_impl = NetworkingImpl::new(
-            session_store.clone(),
-            opened_sessions_tracker,
-            100,
-            50,
-            Duration::from_secs(60),
-            TlsExtensionGetter::default(),
-            #[cfg(feature = "testing")]
-            false,
-        );
-
-        let tag = Tag {
-            session_id,
-            sender: MpcIdentity("party1".to_string()),
-            round_counter: 0,
-        };
-        let session_status = session_store.get(&session_id).unwrap();
-
-        let result = networking_impl.fetch_tx_channel(session_status.value(), &tag);
-
-        // Should return Ok(None) when weak reference is dead
-        assert!(result.is_ok());
-        assert!(result.unwrap().is_none());
-    }
-
-    #[test]
-    fn test_session_status_into_send_value_response() {
-        // Test Completed status
-        let completed = SessionStatus::Completed(Instant::now());
-        let response: SendValueResponse = completed.into();
-        assert_eq!(response.status(), Status::Completed);
-
-        // Test Inactive status
-        let channel_maps = DashMap::new();
-        let inactive = SessionStatus::Inactive((
-            MessageQueueStore::new_uninitialized(channel_maps),
-            Instant::now(),
-        ));
-        let response: SendValueResponse = inactive.into();
-        assert_eq!(response.status(), Status::Inactive);
-
-        // Test Active status (with dead weak reference)
-        let weak: Weak<NetworkSession> = Weak::new();
-        let active = SessionStatus::Active(weak);
-        let response: SendValueResponse = active.into();
-        assert_eq!(response.status(), Status::Active);
-    }
-
-    #[test]
-    fn test_message_queue_store_get_tx_initialized() {
-        let role_1 = Role::indexed_from_one(1);
-        let id_1 = Identity::new("127.0.0.1".to_string(), 6001, Some("party1".to_string()));
-        let mut role_assignment: RoleAssignment<Role> = RoleAssignment::default();
-        role_assignment.insert(role_1, id_1.clone());
-
-        let store =
-            MessageQueueStore::new_initialized(100, &role_assignment, Arc::new(DashMap::new()));
-
-        // Should get Some(tx) for existing identity
-        let result = store.get_tx(&id_1.mpc_identity());
-        assert!(result.is_ok());
-        assert!(result.unwrap().is_some());
-
-        // Should get None for non-existing identity
-        let other_sender = MpcIdentity("party2".to_string());
-        let result = store.get_tx(&other_sender);
-        assert!(result.is_ok());
-        assert!(result.unwrap().is_none());
-    }
-}
